Segregate new trait TNabuLexerNodeChecker from CNabuLexerAbstractRule
